Rename model_checked/ to scheduled/, move remaining shared files into it
"Model checking" implies formal/exhaustive verification, which is
misleading here: this codebase already has a genuinely exhaustive,
TLC-based model-checking concept (.github/workflows/ci-verification.yml),
and this suite's own real-stack scenarios only use
explore_random_interleavings() (sampling), not exhaustive search.

- src/commit_concurrency/model_checked/ -> scheduled/, alongside
  deterministic_scheduler.h, deterministic_scheduler_test.cpp, and
  interleaving_lock_override.h, which now live there too (nothing outside
  this directory uses any of them). This leaves src/commit_concurrency/
  with exactly two peer subdirectories - threaded/ (real OS threads,
  seeded but not exactly replayable) and scheduled/ (single-process,
  cooperative, byte-for-byte replayable) - and no loose top-level files.
- commit_concurrency_model_test -> commit_concurrency_scheduled_test
  (CMake target/binary), commit_concurrency_model ->
  commit_concurrency_scheduled (doctest suite tag).
- rejected_commit_stall.cpp -> rejected_commit.cpp: the old name was
  overly specific to one particular way a rejected commit could go wrong.
- src/commit_concurrency/interleaving.h -> threaded/checkpoint.h (+ its
  test file -> threaded/checkpoint_test.cpp, suite tag "interleaving" ->
  "checkpoint"): this is a distinct, complementary primitive - a manual
  pause/release rendezvous for real OS threads - not part of the
  scheduled/ stack, so "interleavings" was freed up for that instead.
- Comments reworded throughout to describe "systematically exploring the
  interleaving space" (exhaustively where the space is small enough,
  randomly sampling otherwise), rather than "model checking".
