+++ src/commit_concurrency/scheduled/deterministic_scheduler_test.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#include "commit_concurrency/deterministic_scheduler.h"
+#include "commit_concurrency/scheduled/deterministic_scheduler.h"
 
 #define DOCTEST_CONFIG_NO_SHORT_MACRO_NAMES
 #include <algorithm>
