+++ src/consensus/aft/impl/state.h
@@ -4,7 +4,7 @@
 
 #if defined(CCF_STATIC_LIBRARY_BUILD)
 #  error \
-    "consensus/aft/impl/state.h must never be compiled into ccf_kv, ccf_tasks, or ccfcrypto: their compiled objects are linked, unmodified, by every test binary, including one that recompiles this header with a different ccf::pal::Mutex (see src/commit_concurrency/interleaving_lock_override.h) - two different memory layouts for the same class name in one binary would be an ODR violation."
+    "consensus/aft/impl/state.h must never be compiled into ccf_kv, ccf_tasks, or ccfcrypto: their compiled objects are linked, unmodified, by every test binary, including one that recompiles this header with a different ccf::pal::Mutex (see src/commit_concurrency/scheduled/interleaving_lock_override.h) - two different memory layouts for the same class name in one binary would be an ODR violation."
 #endif
 
 #include "ccf/crypto/verifier.h"
